chore(build): drop use of github archive for `build-prod`
Building production from a downloaded archive rather than the Docker
build context was done to guard against the risk that an unclean
working copy could cause WIP code to be deployed.

Today, responsibility for building and deployment are entirely handled
by CD process (Github Actions). This is the best practice and has made
the archive download safeguard unnecessary.

Change the `build-prod` target to use the Docker buildcontext rather
than a downloaded copy of the repository. This is more conventional,
respects the `.dockerignore` file, and removes a difference between
the `build-dev` and `build-prod` targets.

Note also: the `poetry.lock` and `pyproject.toml` files are copied to
`$PYSETUP_PATH` (`/opt/pysetup`) rather that `/opt`. This is what
`build-dev` has done and there's no reason for `build-prod` to do
differently.