chore(build): merge build-dev and build-prod stages
The only remaining difference between the two targets
are the `--no-root` and `--without dev`. The former
is used without incident by both in Courtlistener and
the later can be expressed without different stages by
using the `$BUILD_ENV` argument during construction of
the `poetry install` command given to the `RUN`
instruction.